common: Move repositories to src
