+++ modules/common/src/repositories/item.test.ts
@@ -1,7 +1,7 @@
 import { beforeAll, describe, expect, test } from "vitest";
 import { initializeTestEnvironment } from "@firebase/rules-unit-testing";
 import type { Firestore } from "firebase/firestore";
-import firebasejson from "../firebase.json";
+import firebasejson from "../../firebase.json";
 import type { WithId } from "../lib/typeguard";
 import { ItemEntity } from "../models/item";
